+++ Assets/DeltaDNA/Helpers/Settings.cs
@@ -18,7 +18,7 @@ namespace DeltaDNA
 {
     public class Settings
     {
-        internal static readonly string SDK_VERSION = "Unity SDK v4.8.1";
+        internal static readonly string SDK_VERSION = "Unity SDK v4.9.0";
 
         internal static readonly string ENGAGE_API_VERSION = "4";