+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [VERSION](https://github.com/deltaDNA/unity-sdk/releases/tag/VERSION) (YYYY-MM-DD)
+### Added
+- API for forgetting the user and stopping tracking (GDPR).
+
 ### Fixed
 - Crash in Editor when Android notifications are removed from project.